Basic UI Rough Draft
Finally got a rough draft ported into the VS project that seems to work
with the navbar and with varying screen size. This mostly focuses on the
layout of the website, not the contents.
NOTE: had to remove appsettings.json to get it to buid on my PC. But I
think I can add it back if its absence causes anyone else problems.